+++ lib/multiple_result.dart
@@ -2,3 +2,4 @@ library multiple_result;
 
 export 'src/result.dart';
 export 'src/unit.dart';
+export 'src/async_result.dart';